+++ crates/fuzz/src/traits/instruction_hooks.rs
@@ -1,36 +1,22 @@
-use crate::fuzzing::TridentRng;
-use crate::traits::FuzzClient;
+use crate::trident::Trident;
 
 /// Trait implementing methods for the fuzzed instructions
 pub trait InstructionHooks {
     type IxAccounts;
     /// Set Instruction data
     #[allow(unused_variables)]
-    fn set_data(
-        &mut self,
-        client: &mut impl FuzzClient,
-        fuzz_accounts: &mut Self::IxAccounts,
-        rng: &mut TridentRng,
-    ) {
-    }
+    fn set_data(&mut self, trident: &mut Trident, fuzz_accounts: &mut Self::IxAccounts) {}
 
     /// Set Instruction accounts
     #[allow(unused_variables)]
-    fn set_accounts(
-        &mut self,
-        client: &mut impl FuzzClient,
-        fuzz_accounts: &mut Self::IxAccounts,
-        rng: &mut TridentRng,
-    ) {
-    }
+    fn set_accounts(&mut self, trident: &mut Trident, fuzz_accounts: &mut Self::IxAccounts) {}
 
     /// Set Instruction remaining accounts
     #[allow(unused_variables)]
     fn set_remaining_accounts(
         &mut self,
-        client: &mut impl FuzzClient,
+        trident: &mut Trident,
         fuzz_accounts: &mut Self::IxAccounts,
-        rng: &mut TridentRng,
     ) {
     }
 }