+++ crates/fuzz/src/traits/transaction_setters.rs
@@ -1,6 +1,6 @@
 use super::TransactionHooks;
-use crate::fuzzing::TridentRng;
 use crate::traits::FuzzClient;
+use crate::trident::Trident;
 
 #[doc(hidden)]
 pub trait TransactionSetters: TransactionHooks {
@@ -8,11 +8,15 @@ pub trait TransactionSetters: TransactionHooks {
     ///
     /// Creates a new transaction with the necessary data from the fuzzer
     /// and prepares any accounts needed for execution.
-    fn build(
-        client: &mut impl FuzzClient,
-        fuzz_accounts: &mut Self::IxAccounts,
-        rng: &mut TridentRng,
-    ) -> Self;
+    fn build(trident: &mut Trident, fuzz_accounts: &mut Self::IxAccounts) -> Self
+    where
+        Self: Default,
+    {
+        let mut tx = Self::default();
+        tx.set_instructions(trident, fuzz_accounts);
+
+        tx
+    }
     #[doc(hidden)]
     /// Set accounts before transaction
     fn set_snapshot_before(&mut self, client: &mut impl FuzzClient);
@@ -23,10 +27,5 @@ pub trait TransactionSetters: TransactionHooks {
 
     #[doc(hidden)]
     /// Set instructions for the transaction
-    fn set_instructions(
-        &mut self,
-        client: &mut impl FuzzClient,
-        fuzz_accounts: &mut Self::IxAccounts,
-        rng: &mut TridentRng,
-    );
+    fn set_instructions(&mut self, trident: &mut Trident, fuzz_accounts: &mut Self::IxAccounts);
 }