+++ crates/syn/src/types/trident_fuzz_test_methods.rs
@@ -2,7 +2,5 @@ use syn::Ident;
 
 pub struct TridentFuzzTestMethodsStruct {
     pub ident: Ident,
-    pub client_field: Ident,
-    pub metrics_field: Ident,
-    pub rng_field: Ident,
+    pub trident_field: Ident,
 }