+++ examples/cpi/trident-tests/fuzz_0/instructions/initialize_caller.rs
@@ -44,12 +44,7 @@ pub struct InitializeCallerInstructionData {
 /// Docs: https://ackee.xyz/trident/docs/latest/start-fuzzing/writting-fuzz-test/
 impl InstructionHooks for InitializeCallerInstruction {
     type IxAccounts = FuzzAccounts;
-    fn set_data(
-        &mut self,
-        _client: &mut impl FuzzClient,
-        _fuzz_accounts: &mut Self::IxAccounts,
-        rng: &mut TridentRng,
-    ) {
-        self.data.input = rng.gen_range(0..u16::MAX);
+    fn set_data(&mut self, trident: &mut Trident, _fuzz_accounts: &mut Self::IxAccounts) {
+        self.data.input = trident.gen_range(0..u16::MAX);
     }
 }